requirements.txt
Django==6.0.1
mysqlclient>=2.2.1
python-dotenv
djangorestframework
djangorestframework-simplejwt